Add rules YAML directive
This commit adds the `rules` YAML directive in the Helm Chart.
It is used similarly to `stable/elastalert` [1]

Specifically, the `rules` directive expects a map of <name, rule yaml>
and copies the rules as files to a Kubernetes ConfigMap with the
`yaml` file extension.

Finally, it uses the already there `volume-mount-hack` initContainer
to copy the rule files to the `/opt/elastalert/rules` directory
of the Pod (under `data` volumeMount) and grants RW permissions to
`elasticsearch-server`.

[1]:https://github.com/helm/charts/tree/master/stable/elastalert